#31 [feat] [rename] 구성원 Router 연결 설정에 따라 이름 변경
